Migrate .NET SDK provisioning to Arcade scripts
Replace xaprepare's bespoke `dotnet-install` invocation with Arcade's
standard `eng/common/tools.{sh,ps1}` bootstrap, matching dotnet/sdk,
dotnet/runtime, and dotnet/aspnetcore.

* `global.json`: pin `tools.dotnet` so Arcade's `InitializeDotNetCli`
  knows which SDK to install. darc auto-updates this whenever
  `Microsoft.NET.Sdk` flows from dotnet/dotnet via the existing
  Maestro subscription.
* `eng/install-dotnet.{sh,ps1}`: thin wrappers that set
  `DOTNET_INSTALL_DIR=DOTNET_GLOBAL_INSTALL_DIR=bin/$(Configuration)/dotnet/`
  (preserving the existing install location) and call
  `InitializeDotNetCli` from `eng/common/tools.{sh,ps1}`.
* `Makefile`: `prepare` now depends on a new `install-dotnet` target
  that runs `./eng/install-dotnet.sh` first.
* `build-tools/scripts/PrepareWindows.targets`: add an
  `_InstallDotNet` target that invokes `eng/install-dotnet.ps1`
  before `_BuildXAPrepare`, so `dotnet msbuild Xamarin.Android.sln
  -t:Prepare` (used on Windows CI) is self-bootstrapping.
* `Step_InstallDotNetPreview.cs` is deleted and replaced by
  `Step_PrepareDotNetWorkloads.cs`. The new step assumes the SDK
  is already installed at `bin/$(Configuration)/dotnet/` and only
  performs the Android-specific workload prep:
    * Cleans stale Mono Android runtime/workload NuGet directories.
    * Restores `package-download.proj` (Mono runtime packs +
      Mono/Emscripten workload manifest packages).
    * Copies the workload manifests into the local SDK's
      `sdk-manifests/`.
* Removes obsolete configuration:
    * `Configurables.Urls.DotNetInstallScript` (Unix and Windows)
    * `--dotnet-sdk-archive` xaprepare option and its
      `Context.LocalDotNetSdkArchive` plumbing
    * `DownloadDotNetInstallScript`, `GetInstallationScriptArgs`,
      `InstallDotNetAsync`, `InstallDotNetFromLocalArchiveAsync`
      methods (~150 lines of bespoke install logic).

The SDK install location stays at `bin/$(Configuration)/dotnet/`,
so `dotnet-local.{cmd,sh}` and other consumers continue to work
without changes. CI's `use-dot-net.yaml` is unchanged: it still
provisions a system .NET to bootstrap xaprepare; the pinned preview
SDK install simply moves from xaprepare to Arcade.

Verified locally on Windows: `dotnet msbuild Xamarin.Android.sln
-t:Prepare` after `git clean -xdf bin/Debug/dotnet/` installs the
pinned 11.0.100-preview.5.26268.112 SDK and copies the Mono +
Emscripten workload manifests into `sdk-manifests/`.
