+++ build-tools/xaprepare/xaprepare/Application/Context.cs
@@ -278,11 +278,6 @@ namespace Xamarin.Android.Prepare
 		}
 
 		/// <summary>
-		///   Path to a local .NET SDK archive to use instead of downloading.
-		/// </summary>
-		public string? LocalDotNetSdkArchive { get; set; }
-
-		/// <summary>
 		///   Determines whether or not we are running on a hosted azure pipelines agent.
 		///   These agents have certain limitations, the most pressing being the amount of available storage.
 		///   https://docs.microsoft.com/en-us/azure/devops/pipelines/agents/hosted?view=azure-devops#capabilities-and-limitations.
