+++ build-tools/xaprepare/xaprepare/ConfigAndData/Configurables.Unix.cs
@@ -9,14 +9,5 @@ namespace Xamarin.Android.Prepare
 		{
 			public const string DefaultCompiler = "cc";
 		}
-
-		partial class Urls
-		{
-			// This is the "public" url that we really should be using, but it keeps failing with:
-			// AuthenticationException: The remote certificate is invalid because of errors in the certificate chain: RevocationStatusUnknown
-			// For now we'll grab it directly from GitHub
-			// public static readonly Uri DotNetInstallScript = new Uri ("https://builds.dotnet.microsoft.com/dotnet/scripts/v1/dotnet-install.sh");
-			public static readonly Uri DotNetInstallScript = new Uri ("https://builds.dotnet.microsoft.com/dotnet/scripts/v1/dotnet-install.sh");
-		}
 	}
 }
