+++ build-tools/xaprepare/xaprepare/ConfigAndData/Configurables.Windows.cs
@@ -13,10 +13,5 @@ namespace Xamarin.Android.Prepare
 		partial class Paths
 		{
 		}
-
-		partial class Urls
-		{
-			public static readonly Uri DotNetInstallScript = new Uri ("https://builds.dotnet.microsoft.com/dotnet/scripts/v1/dotnet-install.ps1");
-		}
 	}
 }
