+++ build-tools/xaprepare/xaprepare/Main.cs
@@ -27,7 +27,6 @@ namespace Xamarin.Android.Prepare
 			public string? Configuration       { get; set; }
 			public bool AutoProvision          { get; set; }
 			public bool AutoProvisionUsesSudo  { get; set; }
-			public string? LocalDotNetSdkArchive { get; set; }
 		}
 
 		public static int Main (string[] args)
@@ -98,7 +97,6 @@ namespace Xamarin.Android.Prepare
 				"",
 				{"auto-provision=", $"Automatically install software required by .NET for Android", v => parsedOptions.AutoProvision = ParseBoolean (v)},
 				{"auto-provision-uses-sudo=", $"Allow use of sudo(1) when provisioning", v => parsedOptions.AutoProvisionUsesSudo = ParseBoolean (v)},
-				{"dotnet-sdk-archive=", "Path to a local .NET SDK archive (zip or tar.gz) to use instead of downloading from the internet.", v => parsedOptions.LocalDotNetSdkArchive = v?.Trim () },
 				"",
 				{"h|help", "Show this help message", v => parsedOptions.ShowHelp = true },
 			};
@@ -128,7 +126,6 @@ namespace Xamarin.Android.Prepare
 			Context.Instance.DebugFileExtension    = parsedOptions.DebugFileExtension;
 			Context.Instance.AutoProvision         = parsedOptions.AutoProvision;
 			Context.Instance.AutoProvisionUsesSudo = parsedOptions.AutoProvisionUsesSudo;
-			Context.Instance.LocalDotNetSdkArchive = parsedOptions.LocalDotNetSdkArchive;
 
 			if (!String.IsNullOrEmpty (parsedOptions.Configuration))
 				Context.Instance.Configuration = parsedOptions.Configuration!;
