+++ build-tools/xaprepare/xaprepare/Scenarios/Scenario_AndroidTestDependencies.cs
@@ -15,7 +15,7 @@ namespace Xamarin.Android.Prepare
 
 		protected override void AddSteps (Context context)
 		{
-			Steps.Add (new Step_InstallDotNetPreview ());
+			Steps.Add (new Step_PrepareDotNetWorkloads ());
 
 			// disable installation of missing programs...
 			context.SetCondition (KnownConditions.AllowProgramInstallation, false);
