+++ build-tools/xaprepare/xaprepare/Scenarios/Scenario_Standard.cs
@@ -18,7 +18,7 @@ namespace Xamarin.Android.Prepare
 			if (context == null)
 				throw new ArgumentNullException (nameof (context));
 
-			Steps.Add (new Step_InstallDotNetPreview ());
+			Steps.Add (new Step_PrepareDotNetWorkloads ());
 			Steps.Add (new Step_GenerateFiles (atBuildStart: true));
 			Steps.Add (new Step_GenerateCGManifest ());
 		}
