+++ global.json
@@ -2,6 +2,9 @@
   "sdk": {
     "allowPrerelease": true
   },
+  "tools": {
+    "dotnet": "11.0.100-preview.5.26268.112"
+  },
   "msbuild-sdks": {
     "Microsoft.Build.NoTargets": "3.7.134",
     "Microsoft.DotNet.Arcade.Sdk": "11.0.0-beta.26302.103"
